services/ocr/requirements.txt
fastapi
uvicorn
pytesseract
pillow
python-multipart